Save progress RepositoryFolder
